+++ src/main/java/com/nova/cls/data/services/criteria/goods/MaxQuantityCriterion.java
@@ -3,7 +3,7 @@ package com.nova.cls.data.services.criteria.goods;
 public class MaxQuantityCriterion extends GoodsCriterion {
     private static final String SQL = "quantity <= ?";
 
-    public MaxQuantityCriterion(int quantity) {
+    public MaxQuantityCriterion(long quantity) {
         super(SQL, quantity);
     }
 }
