+++ src/main/java/com/nova/cls/data/services/criteria/goods/MinQuantityCriterion.java
@@ -3,7 +3,7 @@ package com.nova.cls.data.services.criteria.goods;
 public class MinQuantityCriterion extends GoodsCriterion {
     private static final String SQL = "quantity >= ?";
 
-    public MinQuantityCriterion(int quantity) {
+    public MinQuantityCriterion(long quantity) {
         super(SQL, quantity);
     }
 }
